src/main/java/com/restaurant/listing/controller/RestaurantController.java
package com.restaurant.listing.controller;

import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.restaurant.listing.dto.RestaurantDTO;
//import com.restaurant.listing.service.RestaurantService;
import com.restaurant.listing.service.RestaurantService;

@RestController
@RequestMapping("/restaurant")
public class RestaurantController {

	@Autowired
	RestaurantService restaurantService;
	
	// fetch all restaurants list 
	@GetMapping("/fetchAllRestaurants")
	public ResponseEntity<List<RestaurantDTO>> fetchAllRestaurants() {
		List<RestaurantDTO> restaurantList = restaurantService.findAllRestaurants();
		return new ResponseEntity<>(restaurantList, HttpStatus.OK);
	}
	
	// add new restaurant 
	@PostMapping(value = "/addRestaurants", produces = "application/json")
	public ResponseEntity<RestaurantDTO> addRestaurant(@RequestBody RestaurantDTO restaurantDTO) {
		RestaurantDTO restaurantAdded = restaurantService.addRestaurantInDB(restaurantDTO);
		return new ResponseEntity<>(restaurantAdded, HttpStatus.CREATED);
	}
	
	// get restaurant by id 
	@GetMapping("fetchById/{id}")
	public ResponseEntity<RestaurantDTO> findRestaurantById(@PathVariable Integer id ) {
		return restaurantService.findRestaurantById(id);
	}
}
